Backend/App_WorkTime-Report_API/requirements.txt
fastapi
uvicorn
requests
python-dotenv
msal